API: Create a new 'fonts' module under the 'config' one
This makes easier to access font-related config options.
