+++ spyder/plugins/help/plugin.py
@@ -15,13 +15,13 @@ import os
 from qtpy.QtCore import Signal
 
 # Local imports
+from spyder.api.config.fonts import SpyderFontType
 from spyder.api.exceptions import SpyderAPIError
 from spyder.api.plugins import Plugins, SpyderDockablePlugin
 from spyder.api.plugin_registration.decorators import (
     on_plugin_available, on_plugin_teardown)
 from spyder.api.translations import _
 from spyder.config.base import get_conf_path
-from spyder.config.fonts import SpyderFontType
 from spyder.plugins.help.confpage import HelpConfigPage
 from spyder.plugins.help.widgets import HelpWidget
