+++ spyder/plugins/history/plugin.py
@@ -12,12 +12,12 @@ Console History Plugin.
 from qtpy.QtCore import Signal
 
 # Local imports
+from spyder.api.config.fonts import SpyderFontType
 from spyder.api.plugins import Plugins, SpyderDockablePlugin
 from spyder.api.plugin_registration.decorators import (
     on_plugin_available, on_plugin_teardown)
 from spyder.api.translations import _
 from spyder.config.base import get_conf_path
-from spyder.config.fonts import SpyderFontType
 from spyder.plugins.history.confpage import HistoryConfigPage
 from spyder.plugins.history.widgets import HistoryWidget
