+++ spyder/plugins/mainmenu/plugin.py
@@ -17,12 +17,12 @@ from typing import Dict, List, Tuple, Optional, Union
 from qtpy.QtGui import QKeySequence
 
 # Local imports
+from spyder.api.config.fonts import SpyderFontType
 from spyder.api.exceptions import SpyderAPIError
 from spyder.api.plugin_registration.registry import PLUGIN_REGISTRY
 from spyder.api.plugins import SpyderPluginV2, SpyderDockablePlugin, Plugins
 from spyder.api.translations import _
-from spyder.api.widgets.menus import MENU_SEPARATOR, SpyderMenu
-from spyder.config.fonts import SpyderFontType
+from spyder.api.widgets.menus import SpyderMenu
 from spyder.plugins.mainmenu.api import ApplicationMenu, ApplicationMenus
 from spyder.utils.qthelpers import set_menu_icons, SpyderAction
