+++ spyder/plugins/onlinehelp/pydoc_patch.py
@@ -24,9 +24,8 @@ import tokenize
 import warnings
 
 # Local imports
-from spyder.api.config.mixins import SpyderFontsMixin
+from spyder.api.config.fonts import SpyderFontsMixin, SpyderFontType
 from spyder.config.base import _, DEV
-from spyder.config.fonts import SpyderFontType
 from spyder.config.gui import is_dark_interface
 from spyder.py3compat import to_text_string
